[Charts] Replace students of x with echarts
